Refactor: replace formatter with dictutil for dropitems_list functionality across multiple modules and remove unused formatter module
